SAVER: Single-cell Analysis Via Expression Recovery

The SAVER package implements SAVER, a gene expression recovery method for single-cell RNA sequencing (scRNA-seq). Borrowing information across all genes and cells, SAVER provides estimates for true expression levels as well as posterior distributions to account for estimation uncertainty. See Huang et al (2018) for more details.

Source

https://github.com/mohuangx/SAVER